django/requirements.txt
Django==1.8
django-crispy-forms==1.6.0
django-haystack==2.4.1
django-registration==2.1
djangorestframework==3.5.1
et-xmlfile==1.0.1
jason==0.1.7
jdcal==1.3
Markdown==2.6.7
openpyxl==2.4.0
Pillow==3.2.0
pytz==2016.4
requests==2.10.0
simplejson==3.8.2
stripe==1.35.0
Whoosh==2.7.4
xlwt==1.1.2